Add comments to exit functions
Also made direction parameters consistent with "dx/dt decreases/increases through zero" comments. For wall1_exit and wall3_exit, the signs were reversed both in the value parameter and the direction parameter relative to the comment (e.g., the code checked for -dx/dt to increase rather than for dx/dt to decrease).

These changes shouldn't affect the output of the model, but should make the code easier to understand.